refactor(web): split timing helpers into own module, accumulate bests in effect
